fix: move formFieldsFragment to components, modify types
