+++ core/app/[locale]/(default)/(auth)/register/page-data.ts
@@ -2,8 +2,8 @@ import { cache } from 'react';
 
 import { getSessionCustomerId } from '~/auth';
 import { client } from '~/client';
-import { FormFieldsFragment } from '~/client/fragments/form-fields';
 import { graphql, VariablesOf } from '~/client/graphql';
+import { FormFieldsFragment } from '~/components/form-fields/fragment';
 import { bypassReCaptcha } from '~/lib/bypass-recaptcha';
 
 const RegisterCustomerQuery = graphql(
