+++ core/app/[locale]/(default)/account/(tabs)/addresses/_components/customer-edit-address.tsx
@@ -1,8 +1,8 @@
 import { getSessionCustomerId } from '~/auth';
 import { client } from '~/client';
-import { FormFieldsFragment } from '~/client/fragments/form-fields';
 import { graphql, ResultOf } from '~/client/graphql';
 import { ExistingResultType } from '~/client/util';
+import { FormFieldsFragment } from '~/components/form-fields/fragment';
 import { bypassReCaptcha } from '~/lib/bypass-recaptcha';
 
 import { getCustomerAddresses } from '../page-data';
