+++ core/components/form-fields/multiline-text.tsx
@@ -1,12 +1,13 @@
+import { FragmentOf } from 'gql.tada';
 import { useTranslations } from 'next-intl';
 import { ChangeEvent } from 'react';
 
 import { Field, FieldControl, FieldLabel, FieldMessage, TextArea } from '~/components/ui/form';
 
-import { CustomerFields } from '../../app/[locale]/(default)/(auth)/register/_components/register-customer-form';
+import { FormFieldsFragment } from './fragment';
 
 type MultilineTextType = Extract<
-  NonNullable<CustomerFields>[number],
+  FragmentOf<typeof FormFieldsFragment>,
   { __typename: 'MultilineTextFormField' }
 >;
